apps/agent/pyproject.toml
[project]
name = "my-dashboard-agent"
version = "0.0.1"
description = "AI Agent for My Dashboard"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "anthropic>=0.40.0",
    "supabase>=2.7.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "feedparser>=6.0.11",
    "python-dotenv>=1.0.1",
    "youtube-transcript-api>=0.6.2",
    "google-api-python-client>=2.149.0",
    "numpy>=1.26.0",
    "apscheduler>=3.10.4",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
